[haproxy] Support interfaces with '-' chars
Sometimes users may be trying to bind on interfaces that have a '-' in
them; Ansible will helpfully change this to a '_', so we must change it
back if we are to refer to the Linux device. This is mostly an issue
when using bridges. Ansible exposes the Linux device name as the
'device' attribute on the interface details.

Change-Id: Ibe49eaf57992f0564d0e37cb7a7ffc891b2777c4