[nova] Do not set api_database for most services
The reason for this is that, if this is set, a RPC version detection
mechanism in nova.compute.rpcapi will throw exceptions because it is
attempting to look at all cells, which is not allowed when you are a
compute service, for example. Only the nova-api and nova-scheduler
services are allowed to touch this database. If this configuration
section is not in the nova configuration, no attempt is made to iterate
over all cells in this fashion.