Configure bridge mappings with neutron_networks
This variable is a special one that sets up lots of networking-related
configuration for a site. The precise values for the bridge names and
external interfaces are set based on the value of this config, which has
the form like:

  neutron_networks:
    physnet1:
      bridge_name: br-em1
      external_interface: em1
    physnet2:
      bridge_name: br-em2
      external_interface: em2

.. and so on. Additionally, a 'patch_interface' is supported:

  neutron_networks
    othernet1:
      bridge_name: br-other1
      patch_interface: br-em1

In this case, br-other1 will be added as an OVS bridge with a peer
bridge of br-em1; the bridges will be logically connected.